sri_lanka/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
llama-index
llama-index-vector-stores-postgres
llama-index-readers-file
pypdf
pdfplumber
pillow
python-multipart
pydantic
python-dotenv
httpx
llama-index-retrievers-bm25
rank-bm25
llama-index-llms-litellm
llama-index-embeddings-litellm
llama-index-embeddings-huggingface
litellm
google-auth
slowapi
beautifulsoup4
boto3
requests
lxml

